suggestions please
couple things: I am getting notified every minute via vibrations of a Nightwatch alert...which I can't find on the Nightwatch app as to what's causing it. I really don't need ANY alerts; I just use it for the glucose on the phone and the watch. Any idea how to turn off all the notifications?

2nd: I am getting a new Samsung Watch 4 with WEAR OS. I used the former watch app with the Tizen OS on Samsung. Any idea how to integrate Nightwatch with Watch 4 or do I need xDrip or another product?

Lastly: Is there a way to show my blood sugar in a widget on my desktop?

Thanks. New here, so appreciate the help